% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{AddModuleScoreScaled}
\alias{AddModuleScoreScaled}
\title{AddModuleScore from Seurat using scaled expression values}
\usage{
AddModuleScoreScaled(
  object,
  features,
  pool = NULL,
  nbin = 20,
  ctrl = 100,
  k = FALSE,
  assay = NULL,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{
Seurat object
}
\description{
AddModuleScore from Seurat using scaled expression values
}
