% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{CalcPCHeuristics}
\alias{CalcPCHeuristics}
\title{Calculate heuristic metrics to determine optimal number of principal components for downstream analyses}
\usage{
CalcPCHeuristics(
  seurat.obj,
  reduction = "pca",
  store.name = "pca_metrics",
  percent.stdev.range = 0.05,
  set.pcs = NA,
  force_tp = FALSE
)
}
\arguments{
\item{seurat.obj}{Seurat object}

\item{percent.stdev.range}{Percent of PCA standard deviation range to use for}

\item{set.pcs}{Number of PCs to use for maximum distance and derivative calculations}

\item{derivs.change.threshold}{Threshold used for second derivative changes}
}
\value{
Returns \code{object} with optimal PC information within the \code{object@misc} slot
}
\description{
Calculate heuristic metrics to determine optimal number of principal components for downstream analyses
}
