% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Cluster}
\alias{Cluster}
\title{Leiden graph-based clustering}
\usage{
Cluster(
  res.param,
  graph,
  seed.param,
  return.params = TRUE,
  return.clusters = FALSE,
  num.iter = 1
)
}
\arguments{
\item{res.param}{Resolution}

\item{graph}{Graph}

\item{seed.param}{Seed}

\item{return.params}{Return cluster-level results}

\item{return.clusters}{Return cell-level results}

\item{num.iter}{Number of leidenbase iterations}
}
\value{
List containing clustering and/or parameter results
}
\description{
leidenbase wrapper function to parse and select results
}
