% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DefineMarkers}
\alias{DefineMarkers}
\title{Define markers with \code{presto} with additional metrics}
\usage{
DefineMarkers(seurat.obj, group)
}
\arguments{
\item{seurat.obj}{Seurat object}

\item{group}{Column in \code{seurat.obj[[]]} to compare}
}
\value{
A data.frame of markers
}
\description{
Define markers with \code{presto} with additional metrics
}
