% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{DetectAmbientRNA}
\alias{DetectAmbientRNA}
\title{Fit models to cell groups}
\usage{
DetectAmbientRNA(data, lineages, types, batches, anno = NULL, fit.n = 50)
}
\arguments{
\item{data}{TP10K matrix genes by cells}

\item{lineages}{cell lineage vector}

\item{types}{cell type vector}

\item{batches}{batch vector}

\item{anno}{named list mapping lineage to types}

\item{fit.n}{number of genes to dit}
}
\value{
list of models
}
\description{
Fit models to cell groups
}
