% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{Downsample}
\alias{Downsample}
\title{Downsample cells evenly across groups}
\usage{
Downsample(
  cells,
  groups,
  quality.vector = NULL,
  total.cells = NULL,
  cells.per.group = NULL
)
}
\arguments{
\item{cells}{vector of cell barcodes}

\item{groups}{vector of cell groups}

\item{quality.vector}{vector of cell quality}

\item{total.cells}{total number of cells to downsample to}

\item{cells.per.group}{total number of cells to downsample from each group}
}
\value{
Vector of downsampled cells
}
\description{
Downsample cells evenly across groups
}
