% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{FitAmbientModel}
\alias{FitAmbientModel}
\title{Fit model}
\usage{
FitAmbientModel(
  data,
  idents,
  batches,
  anno = NULL,
  coefs.use = NULL,
  fit.n = 50,
  fit.cutoff = 2
)
}
\arguments{
\item{data}{TP10K matrix genes by cells}

\item{idents}{identities to fit}

\item{batches}{batches}

\item{coefs.use}{model coefficients to use}

\item{fit.n}{number of genes to fit}

\item{fit.cutoff}{residual cutoff}
}
\value{
list describing model outputs
}
\description{
Fit model
}
