% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Geosketch}
\alias{Geosketch}
\title{Geosketch Seurat object}
\usage{
Geosketch(object, reduction, dims, num.cells)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduced dimensions to use for sketching}

\item{dims}{Number of dimensions to use}

\item{num.cells}{Number of desired cells}
}
\value{
sketch Seurat object downsampled to desired number of cells
}
\description{
See \href{https://github.com/brianhie/geosketch}{geosketch}\cr\cr
Requires geosketch installation and proper reticulate setup
}
