% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{LoessRegression}
\alias{LoessRegression}
\title{Loess regression}
\usage{
LoessRegression(...)
}
\arguments{
\item{...}{Î}
}
\value{
List of model \code{fit}, \code{x} values, and fitted \code{y} values
}
\description{
Loess regression
}
