% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{Log2RmInf}
\alias{Log2RmInf}
\title{Take log2 and remove infinite results}
\usage{
Log2RmInf(x)
}
\arguments{
\item{x}{}
}
\value{
Log2 transformed input vector, \code{Inf = NA}
}
\description{
Take log2 and remove infinite results
}
