% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{NumCellsPerGroup}
\alias{NumCellsPerGroup}
\title{Calculate number of cells to downsample per group}
\usage{
NumCellsPerGroup(groups, total.cells = NULL, cells.per.group = NULL)
}
\arguments{
\item{groups}{vector of groups for each cell barcode}

\item{total.cells}{total number of cells}

\item{cells.per.group}{number of cells to select per group}
}
\value{
Number of cells per group
}
\description{
Calculate number of cells to downsample per group
}
