% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OutlinedDotPlot}
\alias{OutlinedDotPlot}
\title{Modified version of \code{Seurat::Dotplot} with outlined points}
\usage{
OutlinedDotPlot(
  object,
  assay = NULL,
  features,
  cols = c("lightgrey", "blue"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  group.by = NULL,
  split.by = NULL,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA
)
}
\arguments{
\item{object}{Object}

\item{assay}{Assay}

\item{features}{Features to plot}

\item{cols}{cols}

\item{col.min}{col.min}

\item{col.max}{col.max}

\item{dot.min}{dot.min}

\item{dot.scale}{dot.scale}

\item{group.by}{group.by}

\item{split.by}{split.by}

\item{scale.by}{scale.by}

\item{scale.min}{scale.min}

\item{scale.max}{scale.max}
}
\value{
ggplot2 object
}
\description{
See \code{?Seurat::DotPlot}
}
