% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{PlotAmbientRNA}
\alias{PlotAmbientRNA}
\title{Plot Ambient RNA model fits}
\usage{
PlotAmbientRNA(
  u1,
  u2,
  coefs,
  residuals,
  genes_label = NULL,
  genes_fit = NULL,
  label_n = 10,
  residual_cutoff = 2
)
}
\arguments{
\item{u1}{mean expression in-group}

\item{u2}{mean expreession out-group}

\item{coefs}{model coefficients}

\item{residuals}{model residuals}

\item{genes_label}{genes to label}

\item{genes_fit}{genes used for fitting}

\item{label_n}{number of genes to label}

\item{residual_cutoff}{residual cutoff}
}
\value{
ggplot2
}
\description{
Plot Ambient RNA model fits
}
