% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PrepareGraph}
\alias{PrepareGraph}
\title{Create igraph-compatible graph and save in Seurat object}
\usage{
PrepareGraph(object, reduction, dims, knn)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduced dimension slot to pull from}

\item{dims}{Dimensions to use}

\item{knn}{Number of k-nearest neighbors to use}
}
\value{
Seurat object with igraph graph stored in \code{object@misc$westerlund_graph}
}
\description{
Create igraph-compatible graph and save in Seurat object
}
