\name{print.summary.westerlund_test}
\alias{print.summary.westerlund_test}
\title{Print Summary for Westerlund Test}
\description{
Print method for summary objects of class \code{summary.westerlund_test}.
}
\usage{
\method{print}{summary.westerlund_test}(x, ...)
}
\arguments{
  \item{x}{An object of class \code{summary.westerlund_test}.}

  \item{\dots}{Additional arguments passed to or from other methods.}
}
\value{
Invisibly returns the input object \code{x}.
}
\seealso{
\code{\link{summary.westerlund_test}}, \code{\link{westerlund_test}}
}
