% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{ReduceDims}
\alias{ReduceDims}
\title{Process Seurat seurat.object through PCA}
\usage{
ReduceDims(seurat.obj, n.var.features = 2000, n.pcs = 50, remove.genes, ...)
}
\arguments{
\item{seurat.obj}{Seurat object}

\item{n.var.features}{of variable features to use}

\item{n.pcs}{of PCs to use}

\item{remove.genes}{Genes to remove from variable features and PCA}
}
\value{

}
\description{
Process Seurat seurat.object through PCA
}
