% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{RemoveBackgrounds}
\alias{RemoveBackgrounds}
\title{Remove backgrounds}
\usage{
RemoveBackgrounds(outline = FALSE, ...)
}
\arguments{
\item{outline}{Keep plot outline}

\item{...}{}
}
\value{

}
\description{
Remove backgrounds
}
