% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SavePlot}
\alias{SavePlot}
\title{Save png, pdf and plot object}
\usage{
SavePlot(plot, filename, root = "publication", h = 6, w = 6, s = 1)
}
\arguments{
\item{plot}{\code{ggplot2} object to save}

\item{filename}{Filename}

\item{root}{Directories to add to /plots or /data}

\item{h}{Height}

\item{w}{Width}

\item{s}{Scale}
}
\description{
Save png, pdf and plot object
}
