% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambient.R
\name{SelectPoints}
\alias{SelectPoints}
\title{Select points}
\usage{
SelectPoints(
  x,
  y,
  n,
  dir = "both",
  loess = FALSE,
  nbins = 25,
  bin.type = "equal_width",
  exclude = c(0, 0)
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{n}{n}

\item{dir}{direcetionality}

\item{loess}{use Loess}

\item{nbins}{number of bins}

\item{bin.type}{type of bin}

\item{exclude}{exclud}
}
\value{
points vector
}
\description{
Select points
}
