% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SpaceAxisTitles}
\alias{SpaceAxisTitles}
\title{Space axis titles away from plot area}
\usage{
SpaceAxisTitles(scale = 1, ...)
}
\arguments{
\item{scale}{Increase spacing}

\item{...}{}
}
\value{

}
\description{
Space axis titles away from plot area
}
