% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StartFutureLapply}
\alias{StartFutureLapply}
\title{Set Future plan to multisession}
\usage{
StartFutureLapply(num.workers = 8, memory.per.worker = 1000)
}
\arguments{
\item{memory.per.worker}{}
}
\value{

}
\description{
Set Future plan to multisession
}
