\name{summary.westerlund_test}
\alias{summary.westerlund_test}
\title{Summary for Westerlund Test}
\description{
Summary method for objects of class \code{westerlund_test}.
}
\usage{
\method{summary}{westerlund_test}(object, ...)
}
\arguments{
  \item{object}{An object of class \code{westerlund_test}.}
  \item{\dots}{Additional arguments passed to or from other methods.}
}
\value{
An object of class \code{summary.westerlund_test}.
}
