% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{VarToString}
\alias{VarToString}
\title{Convert variable to string}
\usage{
VarToString(variable)
}
\arguments{
\item{variable}{variable to convert to a string name}
}
\value{
Returns a string of variable name
}
\description{
This function turns variable names into a string.
}
