% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_wrappers.R
\name{AddModuleScoreScaled}
\alias{AddModuleScoreScaled}
\title{AddModuleScore from Seurat using scaled expression values}
\usage{
AddModuleScoreScaled(
  object,
  features,
  pool = NULL,
  nbin = 20,
  ctrl = 100,
  k = FALSE,
  assay = NULL,
  name = "Cluster",
  seed = 1,
  search = FALSE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Genes to use}

\item{pool}{Genes to pull from for controls}

\item{nbin}{Number of bins}

\item{ctrl}{Number of control genes}

\item{k}{Feature clusters}

\item{assay}{Assay to use}

\item{name}{Prefix for metadata columns}

\item{seed}{Seed}

\item{search}{Search for gene names}

\item{...}{}
}
\value{
Seurat object
}
\description{
See \code{?Seurat::AddModuleScore}
}
